Adding PFSP_node library, Pool and Pool_multi library (for multi-GPU approach without WS). Changing evaluate files of place. Deletin parameters.h
